requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
openai
langchain
langchain-community
langchain-google-genai
langchain-chroma
langchain-classic
chromadb
pdfplumber
pypdf
python-multipart
python-dotenv
sentence-transformers
streamlit
transformers
torch